+++ CHANGES.rst
@@ -1,5 +1,9 @@
 Version history
 ===============
+**UNRELEASED**
+
+- Added autoincrement to primary key columns to prevent missing field errors.
+  (`#473 <https://github.com/agronholm/sqlacodegen/issues/473>`_; PR by @jtmonroe)
 
 **4.0.3**