+++ CHANGES.rst
@@ -3,7 +3,7 @@ Version history
 
 **UNRELEASED**
 
-- Added autoincrement to primary key columns to prevent missing field errors.
+- Added autoincrement to primary key columns to prevent missing field errors
   (`#473 <https://github.com/agronholm/sqlacodegen/issues/473>`_; PR by @jtmonroe)
 
 **4.0.3**