+++ CHANGES.rst
@@ -1,7 +1,7 @@
 Version history
 ===============
 
-**UNRELEASED**
+**4.0.4*
 
 - Added autoincrement to primary key columns to prevent missing field errors.
   (`#473 <https://github.com/agronholm/sqlacodegen/issues/473>`_; PR by @jtmonroe)